FIX: Juntando funções user
